+++ R/tournament.R
@@ -8,6 +8,8 @@
 #'     \item{\code{team.Visitor}}{The visiting team for the match}
 #'     \item{\code{points.Home}}{Number of points scored by the home team}
 #'     \item{\code{points.Visitor}}{Number of points scored by the visiting team}
+#'     \item{\code{week}}{Week Number}
+#'     \item{\code{half}}{The half of the season in which the match was played}
 #'   }
 #' @examples
 #' data(tournament)
@@ -26,38 +28,31 @@ make_tournament_dataset <- function(seed = NULL)
 
   all.teams <- c("Athletic Armadillos", "Blundering Baboons", "Cunning Cats", "Defense-less Dogs", "Elegant Emus",
                  "Fabulous Frogs", "Gallivanting Gorillas", "Helpless Hyenas")
-  true.elo <- c(1800, 1200, 1700, 1300, 1600, 1550, 1400, 1450)
-  names(true.elo) <- all.teams
+  means <- c(18, 12, 17, 13, 16, 15.5, 14, 14.5)
+  names(means) <- all.teams
 
-  tournament <- expand.grid(team.Home = all.teams, team.Visitor = all.teams)
+  tournament <- expand.grid(team.Home = all.teams, team.Visitor = all.teams, stringsAsFactors = FALSE)
   tournament <- tournament[tournament$team.Home != tournament$team.Visitor, ]
 
-  tournament$elo.Home <- true.elo[tournament$team.Home]
-  tournament$elo.Visitor <- true.elo[tournament$team.Visitor]
+  tournament$points.Home <- vapply(means[tournament$team.Home] + 3, rpois, 0, n = 1)
+  tournament$points.Visitor <- vapply(means[tournament$team.Visitor], rpois, 0, n = 1)
 
-  tournament$adjust.Home <- 200
+  tournament$week <- 0
+  tournament$week[1] <- 1
+  for(i in 2:nrow(tournament))
+  {
+    t1 <- tournament$team.Home[i]
+    t2 <- tournament$team.Visitor[i]
+    idx <- 1:(i-1)
+    wks <- tournament$week[idx]
+    tm1 <- tournament$team.Home[idx]
+    tm2 <- tournament$team.Visitor[idx]
 
-  tournament$p.Home <- elo.prob(tournament$elo.Home + tournament$adjust.Home, tournament$elo.Visitor)
-  tournament$wins.Home <- stats::runif(nrow(tournament)) < tournament$p.Home
+    tournament$week[i] <- min(setdiff(1:14, wks[tm1 %in% c(t1, t2) | tm2 %in% c(t1, t2)]))
+  }
 
-  tournament$elo.Diff <- tournament$elo.Home + tournament$adjust.Home - tournament$elo.Visitor
-
-  tournament$points.Home <- stats::rpois(nrow(tournament), lambda = 10)
-  tournament$points.Visitor <- tournament$points.Home +
-    ifelse(tournament$elo.Diff >= 0 & tournament$wins.Home >  0, pmin(-floor(tournament$elo.Diff / 100), -1), # home team was supposed to win and did
-    ifelse(tournament$elo.Diff >= 0 & tournament$wins.Home == 0, pmax(2 - floor(tournament$elo.Diff / 100),  1), # home team was supposed to win but didn't
-    ifelse(tournament$elo.Diff <  0 & tournament$wins.Home >  0, pmin(2 + floor(tournament$elo.Diff / 100), -1), # visiting team was supposed to win but didn't
-    ifelse(tournament$elo.Diff <  0 & tournament$wins.Home == 0, pmax(-floor(tournament$elo.Diff / 100),  1), NA # visiting team was supposed to win and did
-  ))))
-
-  stopifnot(isSymmetric(matrix(table(tournament$wins.Home, tournament$points.Home > tournament$points.Visitor), nrow = 2)))
-  tournament$elo.Home <- NULL
-  tournament$elo.Visitor <- NULL
-  tournament$elo.Diff <- NULL
-  tournament$p.Home <- NULL
-  tournament$wins.Home <- NULL
-  tournament$adjust.Home <- NULL
-
+  tournament$half <- ifelse(tournament$week < 8, "First Half of Season", "Second Half of Season")
+  tournament <- tournament[order(tournament$week), ]
   rownames(tournament) <- NULL
   attr(tournament, "out.attrs") <- NULL